+++ modules/module-postgres/src/index.ts
@@ -1,3 +1,1 @@
 export * from './module/PostgresModule.js';
-
-export * as pg_utils from './utils/pgwire_utils.js';